Split server.ts contents into separate files by functions
